requirements.txt
numpy
opencv-python-headless
pillow
scipy
streamlit==1.40
streamlit_drawable_canvas==0.8.0